# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: NSA Training Platform API
  version: "1.0.0"
  description: |
    REST/JSON API for the NSA Training Platform (modular monolith).
    Base path: `/api/v1`. All timestamps are UTC (RFC 3339).

    **Roles:** ADMIN, TEACHER, STUDENT (RBAC + ownership checks).

    **Response envelopes**
    - Success: `{ "data": ..., "message": "Success" }`
    - Error: `{ "error": { "code": "SOME_CODE", "message": "...", "details": [] } }`

    **Implementation status:** Phase 16 — system probes, authentication/RBAC,
    academic-core administration, training locations, class-session scheduling,
    role-scoped schedules, batch attendance, corrections, locking, and Student
    temporally correct attendance history/summary with absence-risk warnings
    are live. Practical assessments with evidence, deterministic Student progress,
    ADMIN completion decisions, verifiable PDF certificates, in-app notifications,
    operational CSV reports, and assignment-scoped Teacher workspaces are live.

    **Authentication:** `POST /api/v1/auth/login` returns a short-lived JWT
    access token (use as `Authorization: Bearer <token>`) and sets an
    HttpOnly `nsa_refresh` cookie (path `/api/v1/auth`) used by
    `/auth/refresh` and `/auth/logout`. Refresh tokens rotate on every use;
    reuse of a revoked token revokes the entire session family.
servers:
  - url: http://localhost:8080
    description: Local development (business endpoints are prefixed /api/v1)
...
paths:
...
  /api/v1/teacher/classes:
    get:
      tags: [teacher]
      summary: List classes assigned to the authenticated teacher
      operationId: listTeacherClasses
      responses:
        "200":
          description: Assigned classes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherClassListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/teacher/classes/{classID}:
    parameters:
      - { $ref: "#/components/parameters/ClassID" }
    get:
      tags: [teacher]
      summary: Get an assigned class with its students and competencies
      operationId: getTeacherClass
      responses:
        "200":
          description: Assigned class workspace data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherClassDetailResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (Phase 3). Refresh token is an HttpOnly cookie.
  parameters:
...
    ClassID:
      name: classID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Required role, assignment, or ownership check failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Referenced resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ClassStatus:
      type: string
      enum: [planning, open, in_progress, completed, cancelled, archived]
    EnrollmentStatus:
      type: string
      enum: [enrolled, transferred, completed, withdrawn]
...
    CompetencyCriterion:
      type: object
      required: [id, course_id, code, name, is_required, sequence_no, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        module_id: { type: [string, "null"], format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: [string, "null"] }
        is_required: { type: boolean }
        sequence_no: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Class:
      type: object
      required: [id, course_id, course_code, course_name, class_code, name, start_date, end_date, maximum_students, enrolled_students, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        course_code: { type: string }
        course_name: { type: string }
        class_code: { type: string }
        name: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        maximum_students: { type: integer }
        enrolled_students: { type: integer }
        status: { $ref: "#/components/schemas/ClassStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    TeacherClassListResponse:
      type: object
      required: [data, message]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Class" } }
        message: { type: string }
    TeacherClassDetail:
      type: object
      required: [class, students, competencies]
      properties:
        class: { $ref: "#/components/schemas/Class" }
        students: { type: array, items: { $ref: "#/components/schemas/Enrollment" } }
        competencies: { type: array, items: { $ref: "#/components/schemas/CompetencyCriterion" } }
    TeacherClassDetailResponse:
      type: object
      required: [data, message]
      properties:
        data: { $ref: "#/components/schemas/TeacherClassDetail" }
        message: { type: string }
...
    Enrollment:
      type: object
      required: [id, class_id, student_id, student_code, full_name, status, enrolled_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        class_id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        student_code: { type: string }
        full_name: { type: string }
        status: { $ref: "#/components/schemas/EnrollmentStatus" }
        enrolled_at: { type: string, format: date-time }
        ended_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: STUDENT_NOT_FOUND
            message:
              type: string
              example: Student not found
            details:
              type: array
              items:
                type: object
                additionalProperties: true
              default: []

# Default security for all business endpoints; system endpoints opt out above.
security:
  - bearerAuth: []